requirements.txt
flask==1.0.2
flask-sqlalchemy==2.3.2
mysqlclient==1.4.1
python-dotenv==0.10.1
flask-bcrypt==0.7.1
flask-jwt-extended==3.16.0